Switch from pnp to bundling
pnp works, but every time I look at this repo I have to remember how and
why... bundling isn't ideal since it has to be done before each release,
but it's more similar to everything else I maintain, and the extra
complexity isn't worth it here.

Also add examples of problems with circular dependencies to the readme,
which resolves #1.